get only own images

+++ src/app.py
@@ -53,6 +53,7 @@ app.include_router(
 async def upload_file(
     file: UploadFile = File(...),
     caption: str = Form(""),
+    user: User = Depends(current_active_user),
     session: AsyncSession = Depends(get_async_session)
 ):
     temp_file_path = None
@@ -74,6 +75,7 @@ async def upload_file(
         if upload_result.response_metadata.http_status_code == 200:
 
             post = Post(
+                user_id = user.id,
                 caption = caption,
                 url = upload_result.url,
                 file_type = "video" if file.content_type.startswith("video/") else "image",
@@ -94,10 +96,15 @@ async def upload_file(
 
 @app.get("/feed")
 async def get_feed(
-    session : AsyncSession = Depends(get_async_session)
+    session : AsyncSession = Depends(get_async_session),
+    user : User = Depends(current_active_user)
 ):
     result = await session.execute(select(Post).order_by(Post.created_at.desc()))
     posts = [row[0] for row in result.all()]
+
+    result = await session.execute(select(User))
+    users = [row[0] for row in result.all()]
+    user_dict = {u.id: u.email for u in users}
 
     post_data = []
 
@@ -106,16 +113,19 @@ async def get_feed(
             {
                 "id": str(post.id),
                 "caption": post.caption,
+                "user_id": str(post.user_id),
                 "url": post.url,
                 "file_type": post.file_type,
                 "file_name": post.file_name,
-                "created_at": post.created_at.isoformat()
+                "created_at": post.created_at.isoformat(),
+                "is_owner": post.user_id == user.id,
+                "email": user_dict[post.user_id]
             }
         )
     return {"post":post_data}
 
 @app.delete("/posts/{post_id}")
-async def delete_post(post_id: str, session: AsyncSession=Depends(get_async_session)):
+async def delete_post(post_id: str, session: AsyncSession=Depends(get_async_session), user: User = Depends(current_active_user)):
     try:
         post_uuid=uuid.UUID(post_id)
 
@@ -125,6 +135,9 @@ async def delete_post(post_id: str, session: AsyncSession=Depends(get_async_sess
         if not post:
             raise HTTPException(status_code=404, detail="Post not found")
         
+        if post.user_id != user.id:
+            raise HTTPException(status_code=403, detail="You are not authorized to delete this post")
+
         await session.delete(post)
         await session.commit()
